pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Hello World"
license = "GNU Affero"
authors = ["Nicholas M. Synovic <nicholas.synovic@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/NicholasSynovic/template_python"
repository = "https://github.com/NicholasSynovic/template_python"
keywords = ["Template"]
classifiers = ["Development Status :: 1 - Planning"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
jsonschema = "^4.23.0"
pytest = "^8.3.3"
pymongo = "^4.10.1"
pandas = "^2.2.3"
requests = "^2.32.3"
polars = "^1.12.0"
ibis-framework = {version = "^9.5.0", extras = ["sqlite"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
